@@ -14,39 +14,42 @@ description: A new Flutter Music application.
 version: 0.0.5
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.6.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
+  flutter_localizations:
+    sdk: flutter
 
-  dart_tags: #forked because of some issues about slow tag reading
-    git: git://github.com/Nailik/dart-tags.git
+  #third party:
   shared_preferences_settings: ^1.0.0+1
   folder_picker: ^0.3.0
 
   flutter_isolate: ^1.0.0+11
 
-  audioplayers:
-    git:
-      url: https://github.com/luanpotter/audioplayers.git
-      ref: go-flutter-desktop
+  audioplayers: ^0.14.2
 
   permission_handler: ^4.2.0+hotfix.3
   archive: ^2.0.13
   flutter_ffmpeg: ^0.2.10
 
   fluttertoast: ^3.1.3
-  flutter_speed_dial: ^1.2.5
-  expandable: ^4.1.2
 
   provider: ^3.0.0+1
-  
+
+  url_launcher: ^5.4.2
   # Dart
   moor: ^2.4.2
   # Database for mobile and Desktop
   moor_ffi: ^0.4.0
-  
+
+  easy_localization: ^2.0.2
+
+  #dart:
+  dart_tags: #forked because of some issues about slow tag reading
+    git: git://github.com/Nailik/dart-tags.git
+
 
 dependency_overrides: #sonst wären folder_picker und shared_preferences_settings inkompatibel und flutter_test
   rxdart: ^0.22.2
@@ -76,8 +79,9 @@ flutter:
   assets:
     - asset/images/placeholder.jpg
     - asset/images/artist_placeholder.jpg
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
+    - asset/images/linkedin.png
+    - asset/langs/en-US.json
+    - asset/langs/de-DE.json
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
