@@ -1,3 +1,4 @@
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flymusic/database/moor_database.dart';
 import 'package:flymusic/screens/tabScreens/other/song_item.dart';
@@ -30,7 +31,7 @@ class _ArtistTrackListScreenState extends State<ArtistTrackListScreen> {
               },
             );
           } else {
-            return Text("no data");
+            return Text(tr("no_songs_in_artist", context: context));
           }
         },
       ),
